+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1263,5 +1263,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-18T14:59:12.873Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-19T05:54:54.980Z")
     r"""The date and time the resource was last updated"""