+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKXG019QVTPNPZYNVQN0"
+        "01K5G8MPM7M7PXV94Q4Q2KZ0J7"
     )
     r"""The id of the resource"""