+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -230,7 +230,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K5ENCKRNR22BVKBXRXDPAZWA"
+    id: Optional[str] = "01K5G8MPFVKXH21M3XA7DKETPM"
     r"""The id of the resource"""