+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -127,7 +127,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKXFXAH5BXBW1Z08T6FW"
+        "01K5G8MPM549A6PB1GBY3N7MNR"
     )
     r"""The id of the resource"""