+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKXFZX4SGTTNJHDRF4C3"
+        "01K5G8MPM6310H5NM0RX42C8G5"
     )
     r"""The id of the resource"""