+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -447,7 +447,7 @@ class Twelve(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKSWHJKG9277SMJ4E28Y"
+        "01K5G8MPGEKZAMVJHD82V68CG2"
     )
 
     status: Optional[ToolsStatus] = "live"