+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -472,7 +472,7 @@ class Tools12(BaseModel):
     http: ToolsHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKT1J88RST8HW320MGM7"
+        "01K5G8MPGKX9VB3AKS4PE7S6NV"
     )
 
     status: Optional[StreamRunAgentToolsStatus] = "live"