+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -193,5 +193,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-18T14:59:12.873Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-19T05:54:54.980Z")
     r"""The date and time the resource was last updated"""