+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ENCKXGXNBMN6B8VZHWP4MN"
+        "01K5G8MPM84TH3AV332WFN7ZR9"
     )
     r"""The id of the resource"""